agents/reeva/requirements.txt
langchain
langgraph
langchain-google-genai
asyncpg
python-dotenv
fastapi
uvicorn
